Move `XmlSerializer` to libOPHD project
