+++ OPHD/States/MapViewStateIO.cpp
@@ -16,8 +16,9 @@
 #include "../StructureManager.h"
 #include "../Map/TileMap.h"
 #include "../Map/MapView.h"
-#include "../XmlSerializer.h"
 #include "../Technology/ResearchTracker.h"
+
+#include <libOPHD/XmlSerializer.h>
 
 #include <NAS2D/Utility.h>
 #include <NAS2D/Filesystem.h>
