+++ OPHD/StructureCatalogue.cpp
@@ -4,7 +4,8 @@
 #include "MapObjects/Structures.h"
 #include "MapObjects/StructureType.h"
 #include "IOHelper.h"
-#include "XmlSerializer.h"
+
+#include <libOPHD/XmlSerializer.h>
 
 #include <NAS2D/Utility.h>
 #include <NAS2D/Filesystem.h>
