+++ OPHD/Technology/TechnologyCatalog.cpp
@@ -1,5 +1,6 @@
 #include "TechnologyCatalog.h"
-#include "../XmlSerializer.h"
+
+#include <libOPHD/XmlSerializer.h>
 
 #include <NAS2D/Utility.h>
 #include <NAS2D/Filesystem.h>
