+++ OPHD/ophd.vcxproj
@@ -263,7 +263,6 @@
     <ClCompile Include="UI\TextRender.cpp" />
     <ClCompile Include="UI\TileInspector.cpp" />
     <ClCompile Include="UI\WarehouseInspector.cpp" />
-    <ClCompile Include="XmlSerializer.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Cache.h" />
@@ -398,7 +397,6 @@
     <ClInclude Include="UI\TileInspector.h" />
     <ClInclude Include="UI\WarehouseInspector.h" />
     <ClInclude Include="WindowEventWrapper.h" />
-    <ClInclude Include="XmlSerializer.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="ophd.rc" />
