+++ libOPHD/XmlSerializer.cpp
