+++ libOPHD/XmlSerializer.h
